Update VMX control sizes to 32 bits, and remove redundant code for default controls
Default controls are just the bits that are defined as must-be-1 in capability MSRs, plus whatever controls HVL determines necessary based upon architecture

+++ hvl/include/vmx.h
@@ -29,11 +29,11 @@ typedef void(*exit_handler_t)(struct vcpu_t*);
 
 typedef struct
 {
-    uint64_t pinbased_ctls;
-    uint64_t exit_ctls;
-    uint64_t entry_ctls;
-    uint64_t primary_procbased_ctls; 
-    uint64_t secondary_procbased_ctls;
+    uint32_t pinbased_ctls;
+    uint32_t exit_ctls;
+    uint32_t entry_ctls;
+    uint32_t primary_procbased_ctls; 
+    uint32_t secondary_procbased_ctls;
     
     // States of required control fields 
     struct
